+++ apps/spruce/src/pages/waterfall/useFilters.test.tsx
@@ -1,5 +1,6 @@
 import { MemoryRouter } from "react-router-dom";
 import { renderHook } from "@evg-ui/lib/test_utils";
+import { buildVariants } from "./testData";
 import { BuildVariant, Version } from "./types";
 import { useFilters } from "./useFilters";
 
@@ -216,8 +217,8 @@ describe("useFilters", () => {
             ...waterfall.buildVariants[0],
             builds: [
               {
-                ...waterfall.buildVariants[0].builds[1],
-                tasks: [waterfall.buildVariants[0].builds[1].tasks[1]],
+                ...waterfall.buildVariants[0].builds[0],
+                tasks: [waterfall.buildVariants[0].builds[0].tasks[1]],
               },
             ],
           },
@@ -278,8 +279,8 @@ describe("useFilters", () => {
             ...waterfall.buildVariants[0],
             builds: [
               {
-                ...waterfall.buildVariants[0].builds[1],
-                tasks: [waterfall.buildVariants[0].builds[1].tasks[1]],
+                ...waterfall.buildVariants[0].builds[0],
+                tasks: [waterfall.buildVariants[0].builds[0].tasks[1]],
               },
             ],
           },
@@ -362,7 +363,7 @@ describe("useFilters", () => {
         buildVariants: [
           {
             ...waterfall.buildVariants[0],
-            builds: [waterfall.buildVariants[0].builds[1]],
+            builds: [waterfall.buildVariants[0].builds[0]],
           },
           waterfall.buildVariants[1],
         ],
@@ -480,82 +481,5 @@ const flattenedVersions: Version[] = [
 const waterfall: {
   buildVariants: BuildVariant[];
 } = {
-  buildVariants: [
-    {
-      id: "1",
-      displayName: "BV 1",
-      builds: [
-        {
-          id: "i",
-          tasks: [],
-          version: "a",
-        },
-        {
-          id: "ii",
-          tasks: [
-            {
-              displayName: "Task 20",
-              displayStatusCache: "started",
-              execution: 0,
-              id: "task_20",
-              status: "started",
-            },
-            {
-              displayName: "Task 15",
-              displayStatusCache: "started",
-              execution: 0,
-              id: "task_15",
-              status: "started",
-            },
-          ],
-          version: "b",
-        },
-      ],
-    },
-    {
-      id: "2",
-      displayName: "BV 2",
-      builds: [
-        {
-          id: "ii",
-          tasks: [
-            {
-              displayName: "Task 100",
-              displayStatusCache: "started",
-              execution: 0,
-              id: "task_100",
-              status: "started",
-            },
-          ],
-          version: "b",
-        },
-      ],
-    },
-    {
-      id: "3",
-      displayName: "BV 3",
-      builds: [
-        {
-          id: "iii",
-          tasks: [
-            {
-              displayName: "Task 1",
-              displayStatusCache: "",
-              execution: 0,
-              id: "task_1",
-              status: "success",
-            },
-            {
-              displayName: "Task 2",
-              displayStatusCache: "task-timed-out",
-              execution: 0,
-              id: "task_2",
-              status: "failed",
-            },
-          ],
-          version: "c",
-        },
-      ],
-    },
-  ],
+  buildVariants,
 };
