+++ apps/spruce/src/pages/waterfall/utils.test.ts
@@ -1,5 +1,6 @@
+import { buildVariants } from "./testData";
 import { Version } from "./types";
-import { groupInactiveVersions } from "./utils";
+import { groupBuildVariants, groupInactiveVersions } from "./utils";
 
 describe("groupInactiveVersions", () => {
   it("correctly groups inactive versions", () => {
@@ -7,86 +8,24 @@ describe("groupInactiveVersions", () => {
 
     expect(res).toStrictEqual([
       {
-        inactiveVersions: [
-          {
-            id: "a",
-            author: "sophie.stadler",
-            activated: false,
-            createTime: new Date("2024-09-20T14:56:08Z"),
-            errors: [],
-            message: "bar",
-            requester: "gitter_request",
-            revision: "a",
-            order: 5,
-          },
-        ],
+        inactiveVersions: [versions[0]],
         version: null,
       },
       {
         inactiveVersions: null,
-        version: {
-          id: "b",
-          author: "sophie.stadler",
-          activated: true,
-          createTime: new Date("2024-09-19T14:56:08Z"),
-          errors: [],
-          message: "foo",
-          requester: "gitter_request",
-          revision: "b",
-          order: 4,
-        },
+        version: versions[1],
       },
       {
-        inactiveVersions: [
-          {
-            id: "c",
-            author: "sophie.stadler",
-            activated: false,
-            createTime: new Date("2024-09-19T14:56:08Z"),
-            errors: [],
-            message: "foo",
-            requester: "gitter_request",
-            revision: "c",
-            order: 3,
-          },
-          {
-            id: "d",
-            author: "sophie.stadler",
-            activated: false,
-            createTime: new Date("2024-09-19T14:56:08Z"),
-            errors: [],
-            message: "foo",
-            requester: "gitter_request",
-            revision: "d",
-            order: 2,
-          },
-          {
-            id: "e",
-            author: "sophie.stadler",
-            activated: false,
-            createTime: new Date("2024-09-19T14:56:08Z"),
-            errors: [],
-            message: "foo",
-            requester: "gitter_request",
-            revision: "e",
-            order: 1,
-          },
-        ],
+        inactiveVersions: null,
+        version: versions[2],
+      },
+      {
+        inactiveVersions: [versions[3], versions[4]],
         version: null,
       },
       {
         inactiveVersions: null,
-        version: {
-          id: "f",
-          author: "sophie.stadler",
-          activated: true,
-          createTime: new Date("2024-09-19T14:56:08Z"),
-          errors: [],
-          message: "foo",
-          requester: "gitter_request",
-          revision: "f",
-          order: 0,
-        },
+        version: versions[5],
       },
     ]);
   });
@@ -96,80 +35,28 @@ describe("groupInactiveVersions", () => {
 
     expect(res).toStrictEqual([
       {
-        inactiveVersions: [
-          {
-            id: "a",
-            author: "sophie.stadler",
-            activated: false,
-            createTime: new Date("2024-09-20T14:56:08Z"),
-            errors: [],
-            message: "bar",
-            requester: "gitter_request",
-            revision: "a",
-            order: 5,
-          },
-          {
-            id: "b",
-            author: "sophie.stadler",
-            activated: true,
-            createTime: new Date("2024-09-19T14:56:08Z"),
-            errors: [],
-            message: "foo",
-            requester: "gitter_request",
-            revision: "b",
-            order: 4,
-          },
-          {
-            id: "c",
-            author: "sophie.stadler",
-            activated: false,
-            createTime: new Date("2024-09-19T14:56:08Z"),
-            errors: [],
-            message: "foo",
-            requester: "gitter_request",
-            revision: "c",
-            order: 3,
-          },
-          {
-            id: "d",
-            author: "sophie.stadler",
-            activated: false,
-            createTime: new Date("2024-09-19T14:56:08Z"),
-            errors: [],
-            message: "foo",
-            requester: "gitter_request",
-            revision: "d",
-            order: 2,
-          },
-          {
-            id: "e",
-            author: "sophie.stadler",
-            activated: false,
-            createTime: new Date("2024-09-19T14:56:08Z"),
-            errors: [],
-            message: "foo",
-            requester: "gitter_request",
-            revision: "e",
-            order: 1,
-          },
-        ],
+        inactiveVersions: [versions[0], versions[1]],
         version: null,
       },
       {
         inactiveVersions: null,
-        version: {
-          id: "f",
-          author: "sophie.stadler",
-          activated: true,
-          createTime: new Date("2024-09-19T14:56:08Z"),
-          errors: [],
-          message: "foo",
-          requester: "gitter_request",
-          revision: "f",
-          order: 0,
-        },
+        version: versions[2],
+      },
+      {
+        inactiveVersions: [versions[3], versions[4]],
+        version: null,
+      },
+      {
+        inactiveVersions: null,
+        version: versions[5],
       },
     ]);
+  });
+});
+
+describe("groupBuildVariants", () => {
+  it("correctly groups build variants from versions", () => {
+    expect(groupBuildVariants(versions)).toStrictEqual(buildVariants);
   });
 });
 
@@ -184,6 +71,7 @@ const versions: Version[] = [
     requester: "gitter_request",
     revision: "a",
     order: 5,
+    waterfallBuilds: null,
   },
   {
     id: "b",
@@ -195,17 +83,77 @@ const versions: Version[] = [
     requester: "gitter_request",
     revision: "b",
     order: 4,
+    waterfallBuilds: [
+      {
+        id: "ii",
+        buildVariant: "1",
+        displayName: "BV 1",
+        tasks: [
+          {
+            displayName: "Task 20",
+            displayStatusCache: "started",
+            execution: 0,
+            id: "task_20",
+            status: "started",
+          },
+          {
+            displayName: "Task 15",
+            displayStatusCache: "started",
+            execution: 0,
+            id: "task_15",
+            status: "started",
+          },
+        ],
+      },
+      {
+        id: "ii2",
+        buildVariant: "2",
+        displayName: "BV 2",
+        tasks: [
+          {
+            displayName: "Task 100",
+            displayStatusCache: "started",
+            execution: 0,
+            id: "task_100",
+            status: "started",
+          },
+        ],
+      },
+    ],
   },
   {
     id: "c",
     author: "sophie.stadler",
-    activated: false,
+    activated: true,
     createTime: new Date("2024-09-19T14:56:08Z"),
     errors: [],
     message: "foo",
     requester: "gitter_request",
     revision: "c",
     order: 3,
+    waterfallBuilds: [
+      {
+        id: "iii",
+        buildVariant: "3",
+        displayName: "BV 3",
+        tasks: [
+          {
+            displayName: "Task 1",
+            displayStatusCache: "",
+            execution: 0,
+            id: "task_1",
+            status: "success",
+          },
+          {
+            displayName: "Task 2",
+            displayStatusCache: "task-timed-out",
+            execution: 0,
+            id: "task_2",
+            status: "failed",
+          },
+        ],
+      },
+    ],
   },
   {
     id: "d",
@@ -239,5 +187,13 @@ const versions: Version[] = [
     requester: "gitter_request",
     revision: "f",
     order: 0,
+    waterfallBuilds: [
+      {
+        id: "i",
+        buildVariant: "1",
+        displayName: "BV 1",
+        tasks: [],
+      },
+    ],
   },
 ];
