+++ apps/spruce/src/pages/waterfall/utils.ts
@@ -1,4 +1,5 @@
-import { GroupedVersion, Version } from "./types";
+import { WaterfallQuery } from "gql/generated/types";
+import { BuildVariant, GroupedVersion, Version } from "./types";
 
 export const groupInactiveVersions = (
   versions: Version[],
@@ -29,4 +30,40 @@ export const groupInactiveVersions = (
   });
 
   return filteredVersions;
+};
+
+export const groupBuildVariants = (
+  versions: WaterfallQuery["waterfall"]["flattenedVersions"],
+): BuildVariant[] => {
+  const bvs: Map<string, BuildVariant> = new Map();
+  versions.forEach(({ activated, id, waterfallBuilds }) => {
+    if (!activated) {
+      return;
+    }
+    waterfallBuilds?.forEach(
+      ({ buildVariant, displayName, id: buildId, tasks }) => {
+        if (!bvs.has(buildVariant)) {
+          bvs.set(buildVariant, {
+            id: buildVariant,
+            displayName,
+            builds: [],
+          });
+        }
+
+        const bv = bvs.get(buildVariant);
+        bv?.builds?.push({
+          id: buildId,
+          version: id,
+          tasks: tasks ?? [],
+        });
+      },
+    );
+  });
+
+  // Although each version's build variants are sorted, we need to make sure the whole list is sorted once combined.
+  const arr: BuildVariant[] = Array.from(bvs.values()).sort((a, b) =>
+    a.displayName.localeCompare(b.displayName),
+  );
+
+  return arr;
 };
